refactor(player): 将玩家列表逻辑迁移至应用服务层
在 CorePlayerService 中实现 PlayerListService，将列表解析与命令捕获从 Tauri 命令层下沉至应用层，并新增 PlayerListError 契约错误类型。
