+++ crates/interface/src/error.rs
@@ -432,6 +432,35 @@ impl std::fmt::Display for OnlineTunnelServiceError {
 
 impl std::error::Error for OnlineTunnelServiceError {}
 
+/// 玩家列表查询失败的契约错误类别。
+///
+/// 用于在线玩家、白名单、封禁、OP 列表等通过控制台命令捕获的服务。
+#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize)]
+#[serde(rename_all = "snake_case")]
+pub enum PlayerListError {
+    /// 客户端提供的输入不合法（如空服务器 ID）。
+    InvalidInput,
+    /// 服务器未在运行（无法写入 stdin 发送命令）。
+    ServerNotRunning,
+    /// 服务装配层不可用。
+    ServiceUnavailable,
+    /// 命令已发出但未在超时内收到回显。
+    CaptureFailed,
+}
+
+impl std::fmt::Display for PlayerListError {
+    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        formatter.write_str(match self {
+            Self::InvalidInput => "invalid player list input",
+            Self::ServerNotRunning => "server is not running",
+            Self::ServiceUnavailable => "player list service unavailable",
+            Self::CaptureFailed => "command capture failed or timed out",
+        })
+    }
+}
+
+impl std::error::Error for PlayerListError {}
+
 #[cfg(test)]
 mod tests {
     use super::*;
